Added Home Work With Arrays and Lists

+++ src/com/softserve/edu/homework07_Person/Cleaner.java
@@ -3,7 +3,7 @@ package com.softserve.edu.homework07_Person;
 public class Cleaner extends Staff {
 
 	private final String TYPE_PERSON = "Cleaner";
-	public Cleaner(String name, int salary) {
+	public Cleaner(String name, double salary) {
 		super(name, salary);
 		System.out.println("Constructr form Cleaner class.");
 	}
